Rename sample.py to sampling.py
To avoid any confusing with function pymc.sample, which is the imported
from model pymc.sample in __init__.py
