+++ pymc/sampling.py
